+++ backend/src/pageHandlers/loginHandler.js
@@ -5,26 +5,43 @@ export const CorrectLogin = async (req, res) => {
     const usernameResult = await dbHandler.GetUser(req.body.username);
 
     if (!usernameResult.success){
-        return res.status(404).send(usernameResult.reason);
+        return res.json({
+            success: false,
+            message: usernameResult.reason,
+            status: 404}
+        );
     }
 
     if (usernameResult.value.rows.length == 0){
-        return res.status(404).send("Couldn't find username");
+        return res.json({
+            success: false,
+            message: "couldn't find username",
+            status: 404}
+        );
     }
 
     const passwordResult = await dbHandler.GetUserPasswordAndSaltWithUsername(req.body.username);
 
     if (!passwordResult.success){
-        return res.status(404).send(passwordResult.reason);
+        return res.json({
+            success: false,
+            message: usernameResult.reason,
+            status: 404}
+        );
     }
 
     if (passwordResult.value.rows.length == 0){
-        return res.status(404).send("Couldn't find password");
+        return res.json({
+            success: false,
+            message: "Couldn't find password",
+            status: 404}
+        );
     }
 
     if (passwordResult.value.rows[0].value == req.body.password){
-        return res.status(200).json({success: true,
-            role: `${usernameResult.value.rows[0].role}`
+        return res.json({success: true,
+            role: usernameResult.value.rows[0].role,
+            status:200
         });
     }
 
@@ -36,24 +53,41 @@ export const GetSalt = async (res, username) => {
     const usernameResult = await dbHandler.GetUser(username);
 
     if (!usernameResult.success){
-        return res.status(404).send(usernameResult.reason);
+        return res.json({
+            success: false,
+            message: usernameResult.reason,
+            status: 404}
+        );
     }
 
     if (usernameResult.value.rows.length == 0){
-        return res.status(404).send("Couldn't find username");
+        return res.json({
+            success: false,
+            message: "couldn't find username",
+            status: 404}
+        );
     }
 
     const passwordResult = dbHandler.GetUserPasswordAndSaltWithUsername(username);
 
     if (!passwordResult.success){
-        return res.status(404).send(passwordResult.reason);
+        return res.json({
+            success: false,
+            message: usernameResult.reason,
+            status: 404}
+        );
     }
 
     if (passwordResult.value.rows.length == 0){
-        return res.status(404).send("Couldn't find password");
+        return res.json({
+            success: false,
+            message: "Couldn't find password",
+            status: 404}
+        );
     }
 
-    return res.status(200).json({success: true,
-        salt: (await passwordResult).value.rows[0].salt
+    return res.json({success: true,
+        salt: passwordResult.value.rows[0].salt,
+        status: 200
     });
 }