+++ backend/src/router.js
@@ -3,12 +3,12 @@ import * as loginHandler from "./pageHandlers/loginHandler.js"
 
 
 export const RegisterRoutes = (server) => {
-    server.post("/login", async (req, res) => {
+    server.post("/api/login", async (req, res) => {
        return await loginHandler.CorrectLogin(req, res);
     });
 
 
-    server.get("/login/:username", async (req, res) => {
+    server.get("/api/login/:username", async (req, res) => {
         return await loginHandler.GetSalt(res, req.params.username);
     })
 }