+++ docs/changes.md
@@ -2,6 +2,7 @@
 
 * Installed axios for api calls.
 * changed npm package manager to yarn for lighter load on machine(npm freezes).
+* Installed crypto-js for hashing user password.
 
 # Backend